+++ src/transpyler/transpyler.py
@@ -8,7 +8,7 @@ from .introspection import Introspection
 from .lexer import Lexer
 from .translate.gettext import gettext_for
 from .translate.translate import translate_namespace
-from .runnable import run
+from .runners import run
 
 # Save useful builtin functions
 _compile = _builtins.compile
